impls/sub.rs step 1
step 2 will split out iterator based parts
