+++ src/impls.rs
@@ -1,4 +1,5 @@
 mod add;
 mod index;
 mod intersection;
+mod sub;
 mod union;
